backend/requirements.txt
asgiref==3.11.1
Django==6.0.3
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.17.0
Markdown==3.10.2
sqlparse==0.5.5
